Inserted how to update source code in devcontainer
I think that many developers never develop code inside a devcontainer. They could forget that this approach is basically very similar to the "manual environment." For this reason, I think that having a paragraph to remind the steps necessary to keep their source code aligned with the last repository commit could be helpful.